Add files via upload
This zip file contains an 835 parsing program that can be launched from the desktop no install needed. The DLLs needed to launch and run are self contained within the application's directory. It only parses out the claim summary information from the 835 file. The 'How-To' files and data dictionary files are also included as well as an Excel macro  that will import the parsed files into an excel workbook for further analysis.  I used the code posted at  'Inteligenz /X12Parser ' to create this program a number of years ago.